requirements.txt
Flask
Flask-Cors
requests
py7zr
Pillow